requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
termcolor>=2.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Dev
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
